Massively simplify the RTD ADC-temperature table generation
The old method painstakingly calculated all resistances of a temperature
range in (small) increments using the Callendar Van Dusen equation. Then
searched for the lower and upper bound of all resistances corresponding to
all ADC values and did a linear interpolation to find the real temperature.
This uses a lot of memory and computations.

The new method instead calculates approximated of temperatures for given
resistances using Newton's algorithm. This requires only the memory for
the table and has virtually no memory overhead and is computationally only
marginally more expensive (the approximations converges in ~3 iterations).